+++ Application/Shared/DTOs/RolePermission/RolePermissionDto.cs
@@ -1,10 +1,9 @@
 namespace Application.Shared.DTOs.RolePermission
 {
-    public class RolePermissionDto : BaseDto<long>
+    public class RolePermissionDto : BaseDto<int>
     {
         public int IdRole { get; set; }
-        public long IdPermission { get; set; }
+        public int IdPermission { get; set; }
         public new int? IdUserCreatedAt { get; set; }
-
     }
 }
